Adds initial unittests

+++ include/pixel.hpp
@@ -5,5 +5,5 @@ struct Pixel {
     uint8_t green;
     uint8_t blue;
     uint8_t alpha;
-    uint8_t value;  // value used for sorting
+    int value;  // value used for sorting
 };
